Still have NoneType string argument in google code for hexlify.
