+++ client-py/casperlabs_client/utils.py
@@ -1,7 +1,6 @@
 import base64
-
-import google.protobuf.text_format
 import google.protobuf.json_format
+import google.protobuf.text_format
 
 
 def encode_base64(a: bytes):
@@ -17,3 +16,17 @@ def hexify(o):
 
 def jsonify(o):
     return google.protobuf.json_format.MessageToJson(o)
+
+
+def print_blocks(response, element_name="block"):
+    count = 0
+    for block in response:
+        print(f"------------- {element_name} {count} ---------------")
+        print_block(block)
+        print("-----------------------------------------------------\n")
+        count += 1
+    print("count:", count)
+
+
+def print_block(block):
+    print(hexify(block))
