+++ client-py/manual-test/nodes_down.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
 
 # Shutdown highway network
-pushd ../../hack/docker || exit
+cd ../../hack/docker || exit
+make down
 make node-0/down
 make node-1/down
 make node-2/down
-popd || exit
